Initial auth token test + Setup page tweaks
If auth tokens were saved, they will be tested, and the streamer will be automatically logged in to both their streamer and bot accounts.
I've also changed the Setup page to automatically redirect to the homepage when login completes; this also means to input is required from the streamer when logging in automatically.